fix(db): Resolve integer overflow in feedback table
- Changes the data type of 'conversation_id' and 'ticket_number' in the 'TicketFeedback' model to BigInteger to prevent 'NumericValueOutOfRange' errors.
- Updates the feedback endpoint to correctly retrieve and save the ticket number.
- Ensures large conversation IDs are handled correctly, preventing crashes when saving feedback.

+++ api/models.py
@@ -83,7 +83,7 @@ class TicketFeedback(Base):
     __tablename__ = 'ticket_feedback'
     id = Column(Integer, primary_key=True)
     conversation_id = Column(BigInteger, nullable=False)
-    ticket_number = Column(Integer, nullable=True)
+    ticket_number = Column(BigInteger, nullable=True)
     action_type = Column(Text, nullable=False)  # 'seen', 'dismissed', 'tag_correction'
     feedback_data = Column(Text, nullable=True)  # JSON string with details
     created_at = Column(DateTime, default=datetime.utcnow)
